fix: rename service accounts to serviceaccountsmgmt
