+++ serviceaccountmgmt/README.md
@@ -7,20 +7,20 @@ Service Accounts API allows you to manage credentails for accessing instance SDK
 To install the package to your project use `go get`:
 
 ```shell
-go get github.com/redhat-developer/app-services-sdk-go/serviceaccounts
+go get github.com/redhat-developer/app-services-sdk-go/serviceaccountmgmt
 ```
 
 ## Usage
 
 ### Importing the package
 
-Import the `github.com/redhat-developer/app-services-sdk-go/serviceaccounts/apiv1internal` package into your code:
+Import the `github.com/redhat-developer/app-services-sdk-go/serviceaccountmgmt/apiv1` package into your code:
 
 ```go
 package main
 
 import (
-    "github.com/redhat-developer/app-services-sdk-go/serviceaccounts/apiv1internal"
+    "github.com/redhat-developer/app-services-sdk-go/serviceaccountmgmt/apiv1"
 )
 ```
