+++ serviceaccountmgmt/apiv1/api_client.go
@@ -3,7 +3,7 @@ package serviceaccounts
 import (
 	"github.com/redhat-developer/app-services-sdk-go/core"
 
-	apiv1 "github.com/redhat-developer/app-services-sdk-go/serviceaccounts/apiv1internal/client"
+	apiv1 "github.com/redhat-developer/app-services-sdk-go/serviceaccountmgmt/apiv1/client"
 )
 
 // Config defines the available configuration options
