+++ serviceaccountmgmt/apiv1/client/.openapi-generator/VERSION
